Clean up TypeScript alias package organization
